Ogromny refactor, upiększenie UIElement (przerobienie metod na inline), zmiana działania UIButton (ułatwienie modyfikacji wyglądu + stany + nowe eventy),  przygotowanie menu pod pole tekstowe

+++ Engine.h
@@ -16,7 +16,7 @@ class Engine
 public:
 	static std::shared_ptr<Engine> getInstance();
 
-	int getNumPlayers() const { return m_players.size(); }
+	int getNumPlayers() const { return static_cast<int>(m_players.size()); }
 	sf::Vector2u getWindowSize() const { return m_window.getSize(); }
 
 	void init(uint32_t width, uint32_t height, std::string const& title);
